tenaNts with auth-signup

+++ del.py
@@ -8,19 +8,18 @@ dir_path = str(path.dirname(path.realpath(__file__)))
 def remove_folders(folder_name):
     res = glob.glob(dir_path+'/**/'+folder_name+'/*', recursive=True)
     skips = [str(dir_path) +'/django/']
+    cnt = 0
     for file_path in res:
         for pth in skips:
             if file_path.startswith(pth):
                 continue
-        if file_path.endswith('__pycache__'):
-            file_path = file_path+'/*'
-            sub_res = glob.glob(file_path)
-            for file_path1 in sub_res:
-                os.remove(file_path1)
-        elif not file_path.endswith('__init__.py'):
-            if path.isfile(file_path):
-                os.remove(file_path)
-    print('Folders removed')
+        skip = False
+        if file_path.endswith('__init__.py') and folder_name == 'migrations':
+            skip = True
+        if not skip and path.isfile(file_path):
+            os.remove(file_path)
+            cnt += 1
+    print(f'{cnt} Files removed from folder {folder_name}')
 
 
 def remove_file_by_extension(ext):
@@ -45,10 +44,10 @@ def remove_migrations():
     if path.exists('db.sqlite3'):
         os.remove('db.sqlite3')
 
-remove_migrations()
 remove_folders('tests')
+remove_folders('__pycache__')
 remove_files('tests.py')
 remove_file_by_extension('pyc')
 remove_file_by_extension('po')
+remove_migrations()
 print('done')
-
